Rename gophish-scrapper.go to gophish-scrapper.go
Gophish Data Retrieval Script

This is a Go script that retrieves data on people who clicked a link or submitted data from a Gophish server. It uses the Gophish API to access data and is built using the Go programming language.

Notable features of this script include:

Ability to retrieve data on people who clicked a link or submitted data

 - Easy-to-use command line interface
 - Support for multiple Gophish servers

To use this script, you will need to have Go installed on your system and have access to a Gophish server.